improve group perf improve same char perf

+++ block.go
@@ -5,7 +5,7 @@ import "golang.org/x/exp/constraints"
 type BlockSize = constraints.Unsigned
 
 type Block struct {
-	OriginalSize int64
+	OriginalSize int
 	List         *LinkedList[*Segment]
 	Buffer       []byte
 }
