+++ compress_revert_bad_segments.go
@@ -1,7 +1,7 @@
 package gompressor
 
 // RevertBadSegments is responsible for reverting bad segments.
-func RevertBadSegments(buf []byte, list *LinkedList[*Segment], size int64) []byte {
+func RevertBadSegments(buf []byte, list *LinkedList[*Segment], size int) []byte {
 	orderedSegments := sortAndFilterSegments(list, false, func(cur *ListEntry[*Segment]) bool {
 		if cur.Value.GetCompressionGains() <= 0 {
 			cur.Remove()
@@ -11,7 +11,7 @@ func RevertBadSegments(buf []byte, list *LinkedList[*Segment], size int64) []byt
 	})
 	for _, entry := range orderedSegments {
 		cur, pos := entry.Segment, entry.Pos
-		if pos < int64(len(buf)) {
+		if pos < int(len(buf)) {
 			panic("decompression should be linear")
 		}
 		buf = append(buf, cur.Decompress()...)
