+++ segment_encode_test.go
@@ -11,12 +11,12 @@ func Test_SegmentEncoding(t *testing.T) {
 		segment := &Segment{
 			Type:   TypeUncompressed,
 			Buffer: []byte{1, 2, 3},
-			Pos:    []int64{1, 2, 3},
+			Pos:    []int{1, 2, 3},
 		}
 		buffer := segment.Encode()
 
 		got, pos := DecodeSegment(buffer)
-		if pos != int64(len(buffer)) {
+		if pos != int(len(buffer)) {
 			t.Fatalf("decode returned wrong buffer position")
 		}
 		require.Equal(t, segment, got)
@@ -26,14 +26,14 @@ func Test_SegmentEncoding(t *testing.T) {
 		segment := &Segment{
 			Type:   TypeRepeatSameChar,
 			Buffer: []byte{1, 2, 3},
-			Pos:    []int64{1, 2, 3},
+			Pos:    []int{1, 2, 3},
 			Repeat: 2,
 		}
 
 		buffer := segment.Encode()
 
 		got, pos := DecodeSegment(buffer)
-		if pos != int64(len(buffer)) {
+		if pos != int(len(buffer)) {
 			t.Fatalf("decode returned wrong buffer position")
 		}
 		require.Equal(t, segment, got)
